refactor(editor): 정보 전달 패널 뷰 분리
